Add bypassDiscovery: false
See OCPNAS-105 for the full context. But basically since by default
bypassdiscovery is true, it means that you can *only* use the devices
listed in the localDevicePath array. Since the UI uses /dev/dm-0 we
set it to false, so we can use both /dev/dm-0 and /dev/disk/by-id/